pyproject.toml
[project]
name = "mAIgic-assistant"
version = "0.1.0"
description = "mAIgic Assistant - AI conversation system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
]

# System dependencies for PyAudio
[project.system-dependencies]
pyaudio = ["portaudio19-dev"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "build"
]
openai = [
    "aiohttp>=3.8.0",
    "websockets>=11.0.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Tell hatchling to include the components from src/
[tool.hatch.build.targets.wheel]
packages = ["src/speech_api", "src/speech_openai_impl"]


[tool.pytest.ini_options]
testpaths = ["src", "tests"]  # Test both components and integration
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long

[tool.mypy]
python_version = "3.12"
warn_return_any = true
strict = true
ignore_missing_imports = true
